ath_0.0.1/requirements.txt
# Web framework
Flask>=2.0.0
flask-socketio>=5.0.0
flask_cors

# File system monitoring
watchdog>=2.1.0

# Data handling
pandas>=1.3.0

# Weather API client
openmeteo-requests>=1.0.0
requests-cache>=0.9.0
retry-requests>=2.0.0

# Timezone handling
timezonefinder>=6.0.0
pytz>=2021.1

# Task scheduling
APScheduler>=3.9.1

# Image handling for icon generation
Pillow>=9.0.0

# Spotify API client
spotipy>=2.19.0

# Cryptography
cryptography>=3.4.7

# Keyboard and hotkey handling
pynput>=1.7.6

# Optional: For development
pytest>=7.0.0
flake8>=4.0.0

# WebSocket support
eventlet>=0.33.0
